Revert "fix ordered map hash function"
This reverts commit 86c297735ead30740af960862b66490b6387826c.

+++ src/ordered_map.h
@@ -32,7 +32,7 @@ public:
 
   bool equals(const Value &other) const;
 
-  uint64_t hash(uint64_t seed) const;
+  unsigned int hash(uint64_t seed) const;
 };
 
 class OrderedMapEntries {
